JS: Split BinaryNomsWriter into Node+Browser
This is because on Node we can write UTF8 directly to a buffer
