+++ js/src/put-cache.js
@@ -42,9 +42,7 @@ export default class OrderedPutCache {
   }
 
   _init(): Promise<DbCollection> {
-    // invariant(false === true);
     return makeTempDir().then((dir): Promise<DbCollection> => {
-      // console.log('creating', dir);
       this._folder = dir;
       const coll = new DbCollection(dir);
       return coll.ensureIndex({hash: 1}, {unique: true}).then(() => coll);
@@ -158,9 +156,13 @@ class DbCollection {
   insert(item: ChunkItem, options: Object = {}): Promise<number> {
     return new Promise((resolve, reject) => {
       options.w = 1;
-      //$FlowIssue
-      const data = new Binary(new Buffer(item.data.buffer));
-      this._coll.insert({hash: item.hash, data: data}, options, (err, result) => {
+      const {buffer, byteOffset, byteLength} = item.data;
+      const ua = new Uint8Array(buffer.slice(byteOffset, byteOffset + byteLength));
+      // $FlowIssue
+      const buf = new Buffer(ua);
+      const data = new Binary(buf);
+
+      this._coll.insert({hash: item.hash, data}, options, (err, result) => {
         if (err) {
           reject(err);
         } else {
